added extra security

+++ routes/api/scheduling.js
@@ -25,32 +25,33 @@ router.post("/add", function(req, res){
         shift:shift,
         sub:sub
     };
-
-    //Makes sure the name does not contain swears
-    if (filter.isProfane(name)) {
-        res.json("Please make sure that your name does not include swears/profane language.");
-    }
-    else {
-        MongoClient.connect(url, function(err, db) {
-            var dbo = db.db(db_name);
-    
-            //Checks if exists
-            dbo.collection(date).findOne({shift:shift, job:job}, function(err, results) {
-                //Job is taken
-                if (results) {/*do nothing*/}
-                //Job is not taken
-                else {
-                    MongoClient.connect(url, function(err, db) {
-                        var dbo = db.db(db_name);
-                        //Inserts into database
-                        dbo.collection(date).insertOne(data);
-                        db.close();
-                    }); 
-                }
+    if(req.user) {
+        //Makes sure the name does not contain swears
+        if (filter.isProfane(name)) {
+            res.json("Please make sure that your name does not include swears/profane language.");
+        }
+        else {
+            MongoClient.connect(url, function(err, db) {
+                var dbo = db.db(db_name);
+        
+                //Checks if exists
+                dbo.collection(date).findOne({shift:shift, job:job}, function(err, results) {
+                    //Job is taken
+                    if (results) {/*do nothing*/}
+                    //Job is not taken
+                    else {
+                        MongoClient.connect(url, function(err, db) {
+                            var dbo = db.db(db_name);
+                            //Inserts into database
+                            dbo.collection(date).insertOne(data);
+                            db.close();
+                        }); 
+                    }
+                });
+                res.json("You have successfully signed up.");
+                db.close();
             });
-            res.json("You have successfully signed up.");
-            db.close();
-        });
+        }
     }
 });
 
@@ -58,24 +59,28 @@ router.post("/add", function(req, res){
 router.post("/get", function(req,res){
     var date = req.body.date;
     var shift = req.body.shift;
-
-    MongoClient.connect(url, function(err, db) {
-        var dbo = db.db(db_name);
-        dbo.collection(date).find({shift:shift}).toArray(function(err, results){
-            res.json(results);
+    if(req.user) {
+        MongoClient.connect(url, function(err, db) {
+            var dbo = db.db(db_name);
+            dbo.collection(date).find({shift:shift}).toArray(function(err, results){
+                res.json(results);
+            });
+            
+            db.close();
         });
-        
-        db.close();
-    });
+    }
 });
 
 //Deletes a user from the signup
 router.post("/delete", function(req, res) {
-    MongoClient.connect(url, function(err, db) {
-        var dbo = db.db(db_name);
-        dbo.collection(req.body.date).deleteOne({job:req.body.job, shift:req.body.shift});
-        res.json("Successfully deleted.");
-      });
+    //Makes sure user in logged in
+    if(req.user) {
+        MongoClient.connect(url, function(err, db) {
+            var dbo = db.db(db_name);
+            dbo.collection(req.body.date).deleteOne({job:req.body.job, shift:req.body.shift});
+            res.json("Successfully deleted.");
+        });
+    }
 });
 
 module.exports = router;
